perf(YfmNote): reduce number of yfm-note tooltip re-renders
