+++ packages/editor/src/extensions/yfm/YfmNote/plugins/YfmNoteTooltipPlugin/index.tsx
@@ -1,48 +1,13 @@
-import {CircleInfo, TrashBin} from '@gravity-ui/icons';
-import type {Node, Schema} from 'prosemirror-model';
-import {Plugin, type Transaction} from 'prosemirror-state';
-import type {EditorView} from 'prosemirror-view';
+import {Plugin} from 'prosemirror-state';
 
-import type {ExtensionDeps} from '../../../../../core';
-import {i18n} from '../../../../../i18n/yfm-note';
-import {BaseTooltipPluginView} from '../../../../../plugins/BaseTooltip';
-import {Toolbar, ToolbarDataType} from '../../../../../toolbar';
-import {removeNode} from '../../../../../utils/remove-node';
-import {noteType} from '../../../index';
+import type {ExtensionDeps} from '#core';
+import {BaseTooltipPluginView} from 'src/plugins/BaseTooltip';
 
-import './index.scss';
+import {noteType} from '../../utils';
 
-enum YfmNoteType {
-    tip = 'tip',
-    info = 'info',
-    warning = 'warning',
-    alert = 'alert',
-}
+import {YfmNoteToolbar} from './YfmNoteToolbar';
 
-const isEnable = () => true;
-const isActive = () => false;
-
-const changeType: (
-    type: YfmNoteType,
-) => (params: {
-    schema: Schema;
-    node: Node;
-    pos: number;
-    dispatch: EditorView['dispatch'];
-    tr: Transaction;
-}) => void =
-    (type) =>
-    ({node, pos, dispatch, tr}) => {
-        dispatch(
-            tr.setNodeMarkup(pos, null, {
-                ...node.attrs,
-                class: `yfm-note yfm-accent-${type}`,
-                'note-type': type,
-            }),
-        );
-    };
-
-export const yfmNoteTooltipPlugin = ({actions, schema}: ExtensionDeps) =>
+export const yfmNoteTooltipPlugin = (_deps: ExtensionDeps) =>
     new Plugin({
         view(view) {
             return new BaseTooltipPluginView(view, {
@@ -50,54 +15,7 @@ export const yfmNoteTooltipPlugin = ({actions, schema}: ExtensionDeps) =>
                 nodeType: noteType(view.state.schema),
                 popupPlacement: ['bottom', 'top'],
                 content: (view, {node, pos}) => (
-                    <Toolbar
-                        editor={actions}
-                        focus={() => view.focus()}
-                        // the yfm class allows to access css variables
-                        // https://github.com/diplodoc-platform/transform/blob/master/src/scss/_common.scss#L17
-                        className="yfm g-md-yfm-note-toolbar"
-                        qa="g-md-toolbar-yfm-note"
-                        data={[
-                            [
-                                YfmNoteType.info,
-                                YfmNoteType.tip,
-                                YfmNoteType.warning,
-                                YfmNoteType.alert,
-                            ].map((type) => ({
-                                id: `note-type-${type}`,
-                                icon: {data: CircleInfo},
-                                title: i18n(type),
-                                type: ToolbarDataType.SingleButton,
-                                isActive,
-                                isEnable,
-                                exec: () =>
-                                    changeType(type)({
-                                        schema,
-                                        pos: pos,
-                                        node: node,
-                                        tr: view.state.tr,
-                                        dispatch: view.dispatch.bind(view),
-                                    }),
-                            })),
-                            [
-                                {
-                                    id: 'note-remove',
-                                    icon: {data: TrashBin},
-                                    title: i18n('remove'),
-                                    type: ToolbarDataType.SingleButton,
-                                    isActive,
-                                    isEnable,
-                                    exec: () =>
-                                        removeNode({
-                                            pos: pos,
-                                            node: node,
-                                            tr: view.state.tr,
-                                            dispatch: view.dispatch.bind(view),
-                                        }),
-                                },
-                            ],
-                        ]}
-                    />
+                    <YfmNoteToolbar node={node} pos={pos} editorView={view} />
                 ),
             });
         },
